Updated base.html
Added login and logout	links	to the base template to put everything together. In order to do	this, we have to determine whether the	current user is logged in	or not in order to display the appropriate link for each case.	
The current user is set in	the HttpRequest object by the authentication middleware.	
You can access it with request.user .	
You	will find a	User	object in	the request even if the user is not authenticated.	
A non-authenticated user is set in the request as an instance of AnonymousUser. 
The best way to check whether the current user is authenticated is by	accessing	its read-only attribute ->is_authenticated.