Updated views.py
The login_required decorator	checks whether the	current user is	authenticated.	If the user is authenticated, it	executes the decorated view;	if the user is not authenticated, it redirects the user to the	login URL with	the originally requested URL as a GET parameter named next. By	doing so, the login view	redirects users	to the URL they were trying to access after they successfully log in.