Updated views.py
We	use	the	login_required	decorator	because	users	have	to	be
authenticated	to	edit	their	profile.	In	this	case,	we	are	using	two
model	forms:	UserEditForm	to	store	the	data	of	the	built-in	user	model
and	ProfileEditForm	to	store	the	additional	profile	data	in	the
custom	Profile	model.	To	validate	the	submitted	data,	we	will
execute	the	is_valid()	method	of	both	forms.	If	both	forms	contain
valid	data,	we	will	save	both	forms,	calling	the	save()	method	to
update	the	corresponding	objects	in	the	database.