Added models.py
When	you	have	to	deal	with	user	accounts,	you	will	find	that	the
user	model	of	the	Django	authentication	framework	is	suitable	for
common	cases.	However,	the	user	model	comes	with	very	basic
fields.	You	may	wish	to	extend	the	user	model	to	include	additional
data.	The	best	way	to	do	this	is	by	creating	a	profile	model	that
contains	all	additional	fields	and	a	one-to-one	relationship	with	the
Django	user	model.