Updated views.py
Implemented	an image list view	that	will	handle	both standard browser	requests	and	AJAX requests, including pagination. When the user	initially loads	the	image	list	page, we	will	display	the	first page	of	images.	When	they	scroll	to	the	bottom	of	the	page,	we	load the	following	page	of	items	via	AJAX and	append	it	to	the	bottom	of the main page.