Created decorators.py
We	will	restrict	our	AJAX	views	to	allow	only	requests	generated	via AJAX.	The	Django	request	object	provides	an	is_ajax()	method	that checks	whether	the request	is	being	made	with	XMLHttpRequest,	which means	it	is	an	AJAX	request.	This	value	is	set	in the HTTP_X_REQUESTED_WITH	HTTP	header,	which	is included	in	AJAX	requests	by most	JavaScript	libraries.

The	preceding code is our	custom	ajax_required	decorator. It	defines	a wrap	function	that	returns an HttpResponseBadRequest object	(HTTP 400 code)	if	the	request	is	not	AJAX. Otherwise, it	returns	the	decorated function.